+++ resources/views/desas/index.blade.php
@@ -12,7 +12,7 @@
         <!-- Tombol Kembali ke Profil Desa -->
         <div class="row mb-3">
             <div class="col-12">
-                <a href="{{ route('desas.profile', ['desa_id' => 1]) }}" class="btn btn-primary btn-block btn-sm">Kembali ke Profil Desa</a>
+                <a href="{{ route('desas.profile', ['desa_id' => auth()->user()->desa->desa_id]) }}" class="btn btn-primary btn-block btn-sm">Kembali ke Profil Desa</a>
             </div>
         </div>
 
@@ -21,9 +21,10 @@
             <div class="col-12">
                 <div class="card">
                     <div class="card-body">
-                        <!-- Menambahkan Class 'table-responsive' untuk Tabel agar responsif -->
-                        <div class="table-responsive py-5 pb-4">
-                            {{ $dataTable->table(['class' => 'table table-striped table-bordered', 'width' => '100%']) }}
+                        <div class="table-responsive py-5 pb-4 dropdown_2">
+                            <div class="container-fluid">
+                                {{ $dataTable->table(['width' => '100%']) }}
+                            </div>
                         </div>
                     </div>
                 </div>
@@ -34,6 +35,7 @@
 
 @push('style')
     @include('layouts.includes.datatable_css')
+    <link rel="stylesheet" href="{{ asset('assets/fonts/fontawesome.css') }}">
 @endpush
 
 @push('scripts')