+++ resources/views/partial/nav-builder.blade.php
@@ -114,14 +114,21 @@ $settings = Utility::settings();
                             <span class="dash-mtext custom-weight">{{ __('Dana') }}</span>
                         </a>
                     </li>
-                @if(auth()->user()->desa)
-                    <a class="dash-link" href="{{ route('desas.profile', ['desa_id' => auth()->user()->desa->desa_id]) }}">
-                        <span class="dash-micon"><i class="ti ti-crown"></i></span>
-                        <span class="dash-mtext custom-weight">{{ __('Profile Desa') }}</span>
-                    </a>
-                @else
-                    <p>{{ __('No Desa data available') }}</p>
-                @endif
+                    <li class="dash-item dash-hasmenu {{ request()->is('desas*') ? 'active' : '' }}">
+                        {{-- <a class="dash-link" href="{{ route('desas.profile', ['desa_id' => 1]) }}"> --}}
+                        <a class="dash-link" href="{{ route('desas.profile', ['desa_id' => auth()->user()->desa->desa_id]) }}">
+                            <span class="dash-micon"><i class="ti ti-crown"></i></span>
+                            <span class="dash-mtext custom-weight">{{ __('Profile Desa') }}</span>
+                        </a>
+                    </li>
+                    {{-- @if(auth()->user()->desa)
+                        <a class="dash-link" href="{{ route('desas.profile', ['desa_id' => auth()->user()->desa->desa_id]) }}">
+                            <span class="dash-micon"><i class="ti ti-crown"></i></span>
+                            <span class="dash-mtext custom-weight">{{ __('Profile Desa') }}</span>
+                        </a>
+                    @else
+                        <p>{{ __('No Desa data available') }}</p>
+                    @endif --}}